Networking Test
Constant two way communication between a client (nc localhost 8000) and the server.

When the client disconnects (killall -KILL nc), the server will close automatically

When the client sends a message to the server, a test response is sent back to the client.

memset is used to clear the buffer after each message is received. (It is also used to clear the random characters at the end of the buffer [line 26])